+++ cmd/schemadoc/cmd_mod2.go
@@ -39,6 +39,8 @@ func (command *moduleToMarkdownCommand) Execute(_ []string) error {
 			TemplateName:         command.TemplateFlags.TemplateName,
 			Title:                command.RenderFlags.Title,
 			Description:          command.RenderFlags.Description,
+			SourcePath:           command.RenderFlags.SourcePath,
+			SourceRawPath:        command.RenderFlags.SourceRawPath,
 			TemplatePath:         command.RenderFlags.TemplatePath,
 			WrapWidth:            command.RenderFlags.WrapWidth,
 			ListMarker:           command.RenderFlags.ListMarker,